feat: tablet size with topbar
